Token: Use long instead of int
Since long has more bits, this avoids integer overflows for more inputs. And
because Java does not allow implicit conversions from int to long, lots of code
is also switched to using long (which is a good thing!).

+++ src/uniol/apt/adt/pn/Place.java
@@ -67,7 +67,7 @@ public class Place extends Node {
 	 * <p/>
 	 * @param t the number of tokens on this place.
 	 */
-	public void setInitialToken(int t) {
+	public void setInitialToken(long t) {
 		setInitialToken(new Token(t));
 	}
